refactor Home page + quick train search
