+++ train-tracker/src/styles/App.css
@@ -1,6 +1,7 @@
 .App {
   text-align: center;
   display: flex;
+  flex-direction: column;
   align-items: center;
   min-height: 100vh;
   background-color: #f4f4f9;
@@ -15,6 +16,10 @@
   box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.1);
   background-color: #ffffff;
   border-radius: 8px;
+}
+
+.content {
+  width: 100%;
 }
 
 .heading-box {
@@ -33,63 +38,4 @@
   margin-right: 10px;
 }
 
-.content {
-  display: flex;
-  justify-content: space-between;
-  width: 100%;
-  gap: 1rem;
-}
 
-.search-container {
-  width: 17%;
-  height: 80vh;
-  padding: 1rem;
-  background-color: #f0f0f5;
-  box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.1);
-  border-radius: 8px;
-}
-
-.app-train-list-container {
-  width: 33%;
-  height: 80vh;
-  padding: 1rem;
-  background-color: #f0f0f5;
-  box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.1);
-  border-radius: 8px;
-}
-
-.map-container {
-  flex: 1;
-  width: 50%;
-  height: 80vh;
-  border-radius: 8px;
-  box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.1);
-  overflow: hidden;
-  position: relative;
-}
-
-.Map {
-  width: 100%;
-  height: 100%;
-}
-
-@media (max-width: 768px) {
-  .content {
-    flex-direction: column;
-  }
-
-  .search-container, .app-train-list-container, .map-container {
-    width: 100%;
-    height: auto;
-    margin-bottom: 1rem;
-    overflow: hidden;
-  }
-}
-
-.close-button:hover{
-  cursor: pointer;
-}
-
-.close-button:active{
-  cursor: pointer;
-}
