+++ train-tracker/src/styles/Home.css
@@ -1,0 +1,61 @@
+.home-page {
+    display: flex;
+    justify-content: space-between;
+    width: 100%;
+    gap: 1rem;
+  }
+  
+
+.search-container {
+    width: 20%;
+    height: 80vh;
+    padding: 1rem;
+    background-color: #f0f0f5;
+    box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.1);
+    border-radius: 8px;
+  }
+  
+  .app-train-list-container {
+    width: 30%;
+    height: 80vh;
+    padding: 1rem;
+    background-color: #f0f0f5;
+    box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.1);
+    border-radius: 8px;
+  }
+  
+  .map-container {
+    flex: 1;
+    width: 50%;
+    height: 80vh;
+    border-radius: 8px;
+    box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.1);
+    overflow: hidden;
+    position: relative;
+  }
+  
+  .Map {
+    width: 100%;
+    height: 100%;
+  }
+  
+  @media (max-width: 768px) {
+    .content {
+      flex-direction: column;
+    }
+  
+    .search-container, .app-train-list-container, .map-container {
+      width: 100%;
+      height: auto;
+      margin-bottom: 1rem;
+      overflow: hidden;
+    }
+  }
+  
+  .close-button:hover{
+    cursor: pointer;
+  }
+  
+  .close-button:active{
+    cursor: pointer;
+  }
